Accounted for shared namespace/inherited method calls
